refactor: move api utils in api package outside pages
+ add response status in error toast
+ fix missing onError in useUpdateUser
