+++ packages/tests/jest/api/db.test.ts
@@ -5,7 +5,7 @@
 import { expect } from '@jest/globals';
 import { MongoClient } from 'mongodb';
 import { ObjectId } from 'bson';
-import db from '@packages/utils/api/db';
+import db from '@packages/api/db';
 
 describe('Db utils', () => {
   const originalEnv = process.env;
