+++ pages/api/auth/[...nextauth].ts
@@ -1,49 +1,11 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import NextAuth from 'next-auth';
+import { ObjectId } from 'bson';
 import DiscordProvider from 'next-auth/providers/discord';
 import { MongoDBAdapter } from '@next-auth/mongodb-adapter';
-import db from '@packages/utils/api/db';
-import { Role } from '@packages/data/role';
-import {
-  isSpecialRoleName,
-  PermissionCategory,
-  PermissionPrivilege,
-  Permissions,
-  SpecialRoleName,
-  SpecialRolesParameters,
-} from '@packages/utils/auth';
-import { ObjectId } from 'bson';
-import { isUserWithInfos, User, UserWithInfos } from '@packages/data/user';
-
-const getUserPermissions = async (user: UserWithInfos) => {
-  const userRoleIds: Role[] = await db.find('roles', {
-    _id: { $in: user.roleIds as string[] },
-  });
-  const userPermissions: Permissions = {};
-  userRoleIds.forEach(role => {
-    if (isSpecialRoleName(role.name) && SpecialRolesParameters[role.name].specialPrivilege) {
-      Object.values(PermissionCategory).forEach(category => {
-        userPermissions[category] =
-          SpecialRolesParameters[role.name as SpecialRoleName].specialPrivilege;
-      });
-    } else {
-      (Object.entries(role.permissions) as [PermissionCategory, PermissionPrivilege][]).forEach(
-        ([category, privilege]) => {
-          if (privilege !== undefined) {
-            const actualUserPermissionForCategory =
-              userPermissions[category] ?? PermissionPrivilege.NONE;
-            userPermissions[category] =
-              actualUserPermissionForCategory > privilege
-                ? actualUserPermissionForCategory
-                : privilege;
-          }
-        },
-      );
-    }
-  });
-
-  return userPermissions;
-};
+import { isUserWithInfos, User } from '@packages/data/user';
+import db from '@packages/api/db';
+import { getUserPermissions } from '@packages/api/auth';
 
 export default async function auth(req: NextApiRequest, res: NextApiResponse) {
   const nextAuth = NextAuth(req, res, {
